refactor: move workflow to lib
