+++ lib/workflow/steps/create-admin-role-and-assign-user.ts
@@ -1,4 +1,4 @@
-import { findInTree, isConflictError } from "@/app/workflow/utils";
+import { findInTree, isConflictError } from "@/lib/workflow/utils";
 import { ApiEndpoint } from "@/constants";
 import { LogLevel, StepId, Var } from "@/types";
 import { z } from "zod";
